Correção métodos "findById" para retornar Exception correta e preenchido dados dos serviços na listagem de prestadore

+++ backend/GAServer/src/main/java/br/com/ga/dao/implementations/ServiceTypeDaoImpl.java
@@ -63,7 +63,7 @@ public class ServiceTypeDaoImpl implements IServiceTypeDao {
     }
 
     @Override
-    public int deleteById(long serviceTypeId) throws Exception {
+    public int deleteById(int serviceTypeId) throws Exception {
         if (serviceTypeId <= 0)
             throw new EntityNotFound("Id inválido");
